:recycle: refactor: 예약슬롯 비즈니스로직에서 cacheservice 분리 및 적용
